requirements.txt
Flask==2.2.5
Jinja2==3.1.2
python-dotenv==1.0.1
requests
